+++ LICENSE
@@ -1,4 +1,5 @@
 Copyright (c) 2015, Kamil Kujawinski
+Copyright (c) 2018, Kamil Kujawinski
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
