+++ {{cookiecutter.plugin_name}}/modules/__init__.py
@@ -1,0 +1,1 @@
+from modules.example_command import {{ cookiecutter.plugin_internal_name|replace('_', ' ')|title|replace(' ', '') }}DoCommand
