+++ {{cookiecutter.plugin_name}}/{{cookiecutter.plugin_internal_name|replace('-', '_')}}.py
@@ -1,9 +1,17 @@
-import sublime
-import sublime_plugin
+# -*- coding: utf-8 -*-
+"""Load and Unload all {{cookiecutter.plugin_name}} modules.
 
+This module exports __all__ modules, which Sublime Text needs to know about.
+The list of __all__ exported symbols is defined in modules/__init__.py.
+"""
 
-class {{ cookiecutter.plugin_internal_name|replace('_', ' ')|title|replace(' ', '') }}DoCommand(sublime_plugin.TextCommand):
-    def run(self, edit):
-        s = sublime.load_settings("{{cookiecutter.plugin_name}}.sublime-settings")
-        first_name = s.get('first_name', 'Noname')
-        self.view.insert(edit, 0, "Hi %s!" % first_name)
+try:
+    from .modules import *
+except ValueError:
+    from modules import *
+except ImportError:
+    # Failed to import at least one module. The current solution
+    import sublime
+    sublime.message_dialog(
+        "{{cookiecutter.plugin_name}} failed to reload properly.\n"
+        "Please restart Sublime Text to fix this!")
